requirements.txt
# requirements.txt

numpy>=1.21.0
pyyaml>=6.0
networkx>=2.6.0
scipy>=1.7.0
stable-baselines3>=1.5.0
torch>=1.10.0
hydra-core>=1.1.0
omegaconf>=2.1.0
tensorboard>=2.7.0
matplotlib>=3.4.0
shimmy>=2.0
modules~=1.0.0
fsspec~=2025.5.1
gymnasium~=1.1.1
pandas~=2.3.0